[dcxdoc] - fixed up notes for image /xdid -S
[dcx]
- moved MircTimeToSystemTime() and SystemTimeToMircTime() to misc.cpp
- fixed typo in dcxdatetime.h checking wrong #IFNDEF
- formatting fixes to dcxdatetime.cpp
- added datetime styles longdateformat, shortdateformat, shortdatecenturyformat, timeformat, right, shownone, and updown
- added datetime /xdid -f, -r and -t


/xdid -f [format]
 Element 	 Description
"d" 	The one- or two-digit day.
"dd" 	The two-digit day. Single-digit day values are preceded by a zero.
"ddd" 	The three-character weekday abbreviation.
"dddd" 	The full weekday name.
"h" 	The one- or two-digit hour in 12-hour format.
"hh" 	The two-digit hour in 12-hour format. Single-digit values are preceded by a zero.
"H" 	The one- or two-digit hour in 24-hour format.
"HH" 	The two-digit hour in 24-hour format. Single-digit values are preceded by a zero.
"m" 	The one- or two-digit minute.
"mm" 	The two-digit minute. Single-digit values are preceded by a zero.
"M" 	The one- or two-digit month number.
"MM" 	The two-digit month number. Single-digit values are preceded by a zero.
"MMM" 	The three-character month abbreviation.
"MMMM" 	The full month name.
"t" 	The one-letter AM/PM abbreviation (that is, AM is displayed as "A").
"tt" 	The two-letter AM/PM abbreviation (that is, AM is displayed as "AM").
"yy" 	The last two digits of the year (that is, 1996 would be displayed as "96").
"yyyy" 	The full year (that is, 1996 would be displayed as "1996").

It is acceptable to include extra characters within the format string to produce a more rich display. However, any nonformat characters must be enclosed within single quotes. For example, the format string "'Today is: 'hh':'m':'s ddddMMMdd', 'yyy" would produce output like "Today is: 04:22:31 Tuesday Mar 23, 1996".

Note   A DTP control tracks locale changes when it is using the default format string. If you set a custom format string, it will not be updated in response to locale changes.


/xdid -r [min|nolimit] [max|nolimit]
range

/xdid -t [timestamp|reset]
set time/date
reset only valid for style shownone

git-svn-id: http://dcx.googlecode.com/svn@448 17a06da3-8c20-fa79-550f-e9588e724b77