[dcx_tools.mrc] changed /return to /returnex
[TString]
added an internal SPACE, TAB, COMMA for use by the class.

[general]
changed all code to use the default tstring space token where we can & to use the internal tokens also.
changed all controls to have an improved parseInfoRequest() that uses a TString prop instead of doing lots of gettoks.
changed some to_num() calls to to_int() when the  number was typed as an int

[xpopup]
changed the $xpopup().style to use a switch statement instead of the if's

[statusbar]
fixed $xdid().tooltip using gettok( 3 ) when it should have been gettok( 4 )


git-svn-id: http://dcx.googlecode.com/svn@606 17a06da3-8c20-fa79-550f-e9588e724b77