[general] added a GdiFlush() to resizeBitmap() after a resize. added a GdiFlush() to dcxLoadIcon() & dcxLoadBitmap() changed ChangeHwndIcon() to only search for the +s & +b flags once.
[dialog]
changed `xdialog -l update` & `xdialog -z` to only do a redraw if the dialog has a CLA rule.

[mwindow]
now accepts an hwnd instead of a name for docking.

[directshow]
changed xdid -a to check for a min of 4 args.
added a file exists check to xdid -a
added a lot of extra error messages to xdid -a

[listview]
added a check for atleast 4 args in subitems to xdid -a
changed xdid -t to use the autoSize() function.


git-svn-id: http://dcx.googlecode.com/svn@651 17a06da3-8c20-fa79-550f-e9588e724b77