[general] added CommonMessage() function to DcxControl class to handle WM_xx messages that are common to all controls, like WM_HELP. Changed all controls to use this new function & removed the controls own message handling for these messages, dll about 10k smaller now. changed some showError() calls to be showErrorEx() to avoid the use of a TString object. changed SetupAlphaBlend() to pass on the client rect to DrawParentsBackground().
[dialog]
added showErrorEx() function.
changed some showError() calls to be showErrorEx() to avoid the use of a TString object.

[box]
changed WM_ERASEBKGND to update CLA if any, fixes box in tabs not showing cla controls correctly if tab is not the first. (method needs testing, may change if a better way is found)


git-svn-id: http://dcx.googlecode.com/svn@675 17a06da3-8c20-fa79-550f-e9588e724b77