[general] changed several MapWindowPoints() to MapWindowRect() (when were these changed? I'm sure i had these as this macro alrdy) changed Dcx::unload() to use the GetWindowFont() & SetWindowFont() & DeleteFont() macros. (again I'm sure i had this set previously) rewrote Dcx::errorex() to use a TString instead of a fixed buffer size.
[dialog]
changed /xdialog -V to set the Vista glass offsets as well as the fake ones for pre vista.
changed CreateVistaStyle() to support Vista+ as well as pre vista.
fixed toXml() to avoid allocating a TString object that's never deleted.
changed toXml() to take the TString arg as a const.

[dwmmodule]
changed how the dwm functions are stored & accessed so they are now part of the module & special dcx functions are used to accesses them.
changed all other code to use these new functions instead of the pointers directly.
changed isAero() to be an inline function.


git-svn-id: http://dcx.googlecode.com/svn@986 17a06da3-8c20-fa79-550f-e9588e724b77