[dcxml] made some things into const's
[listview]
changed xLoadListview() & parseText2Item() so that /xdid -a's flags & icon info etc.. is copied to the items being added when the items being added are plain text (subitems are added without flags etc..)
changed massSetItem() so that the higher flag numbers are masked out when lvi.state is set. This stops problems with the changes mentioned above.


git-svn-id: http://dcx.googlecode.com/svn@1003 17a06da3-8c20-fa79-550f-e9588e724b77