[directshow] changed setAlpha() to use CreateHDCBuffer() changed parseCommandRequest() to use an unsigned numtok.
[comboex]
changed ComboExEditProc() to handle WM_GETDLGCODE messages to we can respond to the return key being pressed. (this was previously commented out for some reason?)

[dll]
updated date to 2012

[mirclinker]
should be checking the highword of version dword.

[misc]
changed IsFile() to strip ctrl codes from the name when checking it, allowing ctrl codes to be used in mirc to preserve double spaces.

git-svn-id: https://dcx.googlecode.com/svn/branches/dcxutf@1089 17a06da3-8c20-fa79-550f-e9588e724b77