First attempt at a commit with git, so let me know if this is ok twig.
[general]
renamed our round() function to dcx_round() to avoid name conflicts with built in round() function.
added a debug message when OleInitialize()/CoGetClassObject()/InitCommonControlsEx() fails.
bumped dev version to 24
changed GetSystemColor() to return an error if its unable to get the glass colour.
added visualstudio 2012 & 2013 solution files.

[linker]
changed eval()/iEval()/exec() to only send the message if the strcpyn() worked.

[modules]
moved windows version checking to a new function GetWindowVersion().
added a windows 8+ version check.

[dock]
added some error checking for DcxDock::mIRCDockWinProc()->WM_DRAWITEM so that image info is only used if it was gotten ok.
changed /xdock -r to mask out the scrollbar flags before checking the position flags.

[dialog]
changed $xdialog().glasscolor to return an error when it fails.

[button][comboex][datetime][divider][listview][pager][box][check][edit][line][list][radio][scroll][text]
changed parseControlStyles() to use TEXT("") instead of just "" in the for loop.

[dshow]
added an error message when /xdid -a is unable to get the video window styles.
changed getPosition() to use a UINT64 pos & to check if GetCurrentPosition() failed & return zero if it did.
changed getDuration() to use a UINT64 pos & to check if GetDuration() failed & return zero if it did.

[statusbar][tab]
changed WM_DRAWITEM handling to check that ImageList_GetImageInfo() worked before using the results.

[toolbar]
added a check to /xdid -w for ImageList_GetIconSize() working before using the results, and an error message when it fails.

[web]
added checks for DispGetParam() working before using the data, and an error message when it fails.