[general] added dcx.props file needed by vs2012/vs2013
[tstring]
added some ts_* defines to be used instead of the string functions directly, this allows switching between unicode & multibyte compiles without lots of extra code etc..
added buffer size tracking to allow us to better manage the memory allocations, this lead to lots of other changes to accommodate this.
added 2 buffer allocation functions to be called instead of new, these functions insure the buffer size is a multiple of 16bytes. This size was chosen as it allows the use of some sse2+ functions that can read strlen()+16 bytes in the worst case.
renamed m_pWString as m_pTempString to make its usage clearer.
added size() function to return the buffersize in use.
added mreplace() function to replace all chars within a string that match any char in a supplied string with a supplied char.
changed findtok() to return a size_t