[general] changed all dialog collection functions to check m_vpDialog.empty() first.
[dcxml]
fixed bug in parseStyle() where var named styles was used when it should have been tiStyles
added a C++11 version of the loop to parseId()

[webctrl]
added error message to /xdid -n for when we can't allocate memory.

[cla]
changed getMinMaxInfo() to be a const function.
changed parsePath() to use an unsigned depth var