[general] changed to_WindowStyle() & to_WindowExStyle() to not take a reference. added check for the existence of SwitchObjects.h before use. removed some commented code.
[xpopup]
misc casting changes.

[xswitchflags]
removed some commented code.
changed m_bFalse to be constexpr.
changed the flags allowed to be from A to z including the non-letters between the cases as this simplifies the if() statement with no cost to us.

[line]
fixed bug in /xdid -t command that caused to to always fail.

[listview]
added some default constructors where needed.
changed isListViewStyle() to take a WindowStyle arg.
misc formatting changes.