[calendar] added getCurrentCTime() function. changed getValue() to use getCurrentCTime() added $xdid().border to get the border width of the calendar. added $xdid().count to get the calendar count. added $xdid().sel to get the currently selected ctime. added $xdid().view to get the current view:   zero = Month view   one = Year view   two = decade view   three = century view added $xdid().firstdayofweek which returns the current first day of the week & if thats the os setting.   [1|0] [Monday|Tuesday|Wednesday|Thursday|Friday|Saturday|Sunday] added $xdid().calcolor to return the colours for parts of the dropdown calendar part of the control:   $xdid(dname,id,background).calcolour   $xdid(dname,id,monthbk).calcolour   $xdid(dname,id,text).calcolour   $xdid(dname,id,titlebk).calcolour   $xdid(dname,id,titletext).calcolour   $xdid(dname,id,trailingtext).calcolour I may add an optiion to get all colours at once later...
[datetime]
added $xdid().ideal to return the ideal width & height of the control.
added $xdid().calcolor to return the colours for parts of the dropdown calendar part of the control:
  $xdid(dname,id,background).calcolour
  $xdid(dname,id,monthbk).calcolour
  $xdid(dname,id,text).calcolour
  $xdid(dname,id,titlebk).calcolour
  $xdid(dname,id,titletext).calcolour
  $xdid(dname,id,trailingtext).calcolour
I may add an optiion to get all colours at once later...
added $xdid().calfont to get the details for the font used in the calendar dropdown.
added $xdid().calstyle to get the styles applied to the calendar dropdown.
These will be a combination of daystate,multiselect,weeknumbers,notodaycircle,notoday,notrailingdates,shortdaysofweek,noselchangeonnav
added /xdid -c to allow setting the dropdown calendars colours.
  /xdid dname id [background|monthbk|text|titlebk|titletext|trailingtext] [RGB]
added /xdid -S to allow changing the dropdown calendars styles on the fly
  /xdid -S dname id daystate multiselect weeknumbers notodaycircle notoday shortdaysoftheweek noselchangeonnav